codigo_paginas_formulario_e_DB/requirements.txt
streamlit
pydantic
pydantic[email]
sqlalchemy
python-dotenv
pymysql 
pandas 
numpy
mysql-connector-python
streamlit_authenticator
statsmodels
plotly
matplotlib
pandas
